+++ BambooTracker/configuration.hpp
@@ -160,11 +160,14 @@ public:
 	bool getReflectInstrumentNumberChange() const;
 	void setFixJammingVolume(bool enabled);
 	bool getFixJammingVolume() const;
+	void setMuteHiddenTracks(bool enabled);
+	bool getMuteHiddenTracks() const;
 private:
 	bool warpCursor_, warpAcrossOrders_, showRowNumHex_, showPrevNextOrders_, backupModules_;
 	bool dontSelectOnDoubleClick_, reverseFMVolumeOrder_, moveCursorToRight_, retrieveChannelState_;
 	bool enableTranslation_, showFMDetuneSigned_, fill00ToEffectValue_, moveCursorHScroll_;
 	bool overwriteUnusedUnedited_, writeOnlyUsedSamples_, reflectInstNumChange_, fixJamVol_;
+	bool muteHiddenTracks_;
 
 	// Edit settings
 public: