+++ BambooTracker/gui/configuration_handler.cpp
@@ -182,6 +182,7 @@ bool ConfigurationHandler::saveConfiguration(std::weak_ptr<Configuration> config
 		settings.setValue("writeOnlyUsedSamples",	configLocked->getWriteOnlyUsedSamples());
 		settings.setValue("reflectInstNumChange",	configLocked->getReflectInstrumentNumberChange());
 		settings.setValue("fixJammingVolume",		configLocked->getFixJammingVolume());
+		settings.setValue("muteHiddenTracks",		configLocked->getMuteHiddenTracks());
 		settings.endGroup();
 
 		// Edit settings
@@ -322,6 +323,7 @@ bool ConfigurationHandler::loadConfiguration(std::weak_ptr<Configuration> config
 		configLocked->setWriteOnlyUsedSamples(settings.value("writeOnlyUsedSamples", configLocked->getWriteOnlyUsedSamples()).toBool());
 		configLocked->setReflectInstrumentNumberChange(settings.value("reflectInstNumChange", configLocked->getReflectInstrumentNumberChange()).toBool());
 		configLocked->setFixJammingVolume(settings.value("fixJammingVolume", configLocked->getFixJammingVolume()).toBool());
+		configLocked->setMuteHiddenTracks(settings.value("muteHiddenTracks", configLocked->getMuteHiddenTracks()).toBool());
 		if (settings.contains("autosetInstrument")) {	// For compatibility before v0.4.0
 			configLocked->setInstrumentMask(!settings.value("autosetInstrument").toBool());
 			settings.remove("autosetInstrument");