+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - [#2], [#213], [#214] - ADPCM sample editor (thanks [@freq-mod])
+- [#98] - Track visibility control (thanks [@pedipanol], [@freq-mod])
 
 ### Fixed
 - Resize width of the order list after closing configuration dialog